Fixed: FS#1812 - Empty trash bin removes /var/www if del_status table returns empty web_domain field.
git-svn-id: svn://svn.ispconfig.org/ispconfig/branches/ISPConfig-2.2.stable@1249 418fc7b0-dcfb-0310-bdb1-f3812d3f35a6

+++ install_ispconfig/scripts/lib/config.lib.php
@@ -2720,6 +2720,7 @@ function web_user_clean(){
       switch ($item["doctype_id"]) {
       case 1013:
           //Verzeichnisse löschen
+		  if(trim($item["web_domain"]) != '') {
           $web_path = $item["pfad"];
           if($item["web_host"] == ""){
             $web_path_realname = $mod->system->server_conf["server_path_httpd_root"] ."/". $item["web_domain"];
@@ -2745,16 +2746,19 @@ function web_user_clean(){
 
             $mod->system->delgroup("web".$item['doc_id']);
           }
+		  }
 
           $mod->db->query("DELETE FROM del_status WHERE id = '".$item["id"]."'");
       break;
       case 1014:
+		if(trim($item["pfad"]) != '') {
           //User-Verzeichnis löschen
           if(is_dir($item["pfad"])){
             $mod->log->caselog("rm -fr ".$item["pfad"], $this->FILE, __LINE__);
           }
           $mod->system->deluser($item["name"]);
           $mod->db->query("DELETE FROM del_status WHERE id = '".$item["id"]."'");
+		}
       break;
       }
     }
